requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
    #   openai
    #   starlette
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
distro==1.9.0
    # via openai
fastapi==0.117.1
    # via -r requirements.in
filelock==3.19.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
google-auth==2.40.3
    # via google-genai
google-genai==1.38.0
    # via -r requirements.in
grpcio==1.74.0
    # via qdrant-client
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hf-xet==1.1.10
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   -r requirements.in
    #   google-genai
    #   openai
    #   qdrant-client
huggingface-hub==0.34.4
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via torch
jiter==0.11.0
    # via openai
joblib==1.5.2
    # via scikit-learn
lxml==6.0.1
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.3
    # via
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.3
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.8.90
    # via torch
openai==1.108.1
    # via -r requirements.in
packaging==25.0
    # via
    #   huggingface-hub
    #   pytest
    #   transformers
pillow==11.3.0
    # via sentence-transformers
pluggy==1.6.0
    # via pytest
portalocker==3.2.0
    # via qdrant-client
protobuf==6.32.1
    # via
    #   -r requirements.in
    #   qdrant-client
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
    #   google-genai
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via -r requirements.in
python-dotenv==1.1.1
    # via pydantic-settings
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
qdrant-client==1.15.1
    # via -r requirements.in
regex==2025.9.1
    # via transformers
requests==2.32.5
    # via
    #   -r requirements.in
    #   google-genai
    #   huggingface-hub
    #   transformers
rsa==4.9.1
    # via google-auth
safetensors==0.6.2
    # via transformers
scikit-learn==1.7.2
    # via sentence-transformers
scipy==1.16.2
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.47.3
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via google-genai
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.0
    # via transformers
tomli==2.2.1
    # via -r requirements.in
torch==2.8.0
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.56.1
    # via sentence-transformers
triton==3.4.0
    # via torch
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   google-genai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   starlette
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   qdrant-client
    #   requests
websockets==15.0.1
    # via google-genai

# The following packages are considered to be unsafe in a requirements file:
# setuptools
